app/backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
sqlalchemy>=2.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0
google-generativeai>=0.3.0
numpy>=1.24.0,<2.0  # Pin to <2.0 for ChromaDB compatibility
pandas>=2.0.0
python-multipart>=0.0.6

# Optional: Multi-LLM support
openai>=1.0.0  # For OpenAI provider (optional)
# anthropic>=0.7.0  # For Claude (future)

# Phase 2.2: RAG System
chromadb>=0.4.22  # Vector database
pypdf>=3.17.0  # PDF text extraction
pdfplumber>=0.10.3  # Advanced PDF parsing (tables, layout)

